requirements.txt
aniso8601==7.0.0
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
Flask==1.1.1
flask-blacklist==0.0.1
Flask-Dance==2.2.0
Flask-JWT==0.3.2
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
html5lib==1.0.1
idna==2.8
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
mccabe==0.6.1
mechanize==0.4.2
oauthlib==2.1.0
openapi-spec-validator==0.2.7
pep8==1.7.1
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.0
swagger-ui-bundle==0.0.5
typed-ast==1.4.0
urllib3==1.25.3
URLObject==2.4.3
uWSGI==2.0.18
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.2
